Extract BaseShaderDialect from HintedShaderDialect.
